+++ CHANGELOG.md
@@ -2,7 +2,8 @@
 
 ## Ongoing updates
 
-- PR [405](https://github.com/plugwise/python-plugwise-usb/pull/405): Fix recent Ruff errors
+- PR [409](https://github.com/plugwise/python-plugwise-usb/pull/409): Fix recent Ruff errors
+- PR [405](https://github.com/plugwise/python-plugwise-usb/pull/405): Try adding plus-device pairing (untested!!)
 
 ## v0.47.2 - 2026-01-29